flask-server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-auth-oauthlib==1.2.0
pymongo==4.10.1
python-dotenv==1.0.0